add dto for user
