+++ src/dto/mod.rs
@@ -1,1 +1,1 @@
-pub(crate) mod achievement;
+pub mod user;
